Add simple process intialization and dependencies
Add code to intialize the focused process and the field of processes in
general.
For the focus process, PID and address size are relatively final. User
specification for remaining memory accesses is not yet in place, so
that detail is currently static. The start specification is awaiting
further clarification. The last page spec is a running spec and will be
set and changed elsewhere.
For the process field, PID and address size are relatively final. Rem-
aining memory accesses is currentl generated as a range between 1000 and
11000. The start spec is awaiting further clarification. The last page
spec is a running spec and will be set and changed elsewhere.
The time and stdlib header were added to support random number
generation.